requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
audioread==3.1.0
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
decorator==5.2.1
einops==0.8.2
fastapi==0.133.1
ffmpy==1.0.0
filelock==3.24.3
flatbuffers==25.12.19
fsspec==2026.2.0
gradio==6.7.0
gradio_client==2.2.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
nltk==3.9.3
numba==0.64.0
numpy==2.4.2
onnxruntime==1.24.2
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pillow==12.1.1
platformdirs==4.9.2
pooch==1.9.0
protobuf==7.34.0
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
qwen-tts==0.1.1
regex==2026.2.19
requests==2.32.5
rich==14.3.3
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.10.0
torchaudio==2.10.0
tqdm==4.67.3
transformers==4.57.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
